[Bitmex] Move tests to new packages
