+++ xchange-bitmex/src/test/java/org/knowm/xchange/bitmex/service/trade/BitmexOrderReplaceTest.java
@@ -1,19 +1,22 @@
-package org.knowm.xchange.bitmex.service;
+package org.knowm.xchange.bitmex.service.trade;
 
-import static org.knowm.xchange.bitmex.BitmexPrompt.PERPETUAL;
-
-import java.math.BigDecimal;
-import java.util.Date;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.knowm.xchange.ExchangeFactory;
 import org.knowm.xchange.ExchangeSpecification;
 import org.knowm.xchange.bitmex.BitmexExchange;
+import org.knowm.xchange.bitmex.service.BitmexMarketDataService;
+import org.knowm.xchange.bitmex.service.BitmexTradeService;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.Order;
 import org.knowm.xchange.dto.marketdata.OrderBook;
 import org.knowm.xchange.dto.trade.LimitOrder;
 import org.knowm.xchange.utils.CertHelper;
+
+import java.math.BigDecimal;
+import java.util.Date;
+
+import static org.knowm.xchange.bitmex.BitmexPrompt.PERPETUAL;
 
 /** @author Nikita Belenkiy on 18/05/2018. */
 public class BitmexOrderReplaceTest {
